+++ .env
@@ -1,1 +1,3 @@
-DATABASE_URL = mongodb+srv://solobachi:9172000SoloBachi@cluster0.xnnt9.mongodb.net/solobachi?retryWrites=true&w=majority
+PORT=3000
+DATABASE_URL=mongodb+srv://savi:savi@cluster0.g2gtj.mongodb.net/SAVI?retryWrites=true&w=majority
+SECRET=123456789
