+++ model/existingProposalModel.js
@@ -1,43 +1,63 @@
-const mongoose  = require('mongoose')
-
-const existingProposal = new mongoose.Schema({
-    fullName:{
-      first:String,
-      last:String
-    },
+const mongoose=require('mongoose')
+const structure2=new mongoose.Schema({
+    fullname:{
+        type: String,
+        required:true
+    }, 
     email:{
-    type:String,
-    required:true,
-    unique:true
-   },
-   dateOfBirth:Date,
-   NIN:{
-    type:Number,
-    unique:true,
-    required:true
-   },
-   address:{
-    type:String,
-    lowercase:true,
-    required:true
-   },
-   phoneNumber:{
-    type:Number,
-    unique:true,
-    required:true,
-   },
-   stateOfOrigin:{
-    type:String,
-    required:true
-   },
-   productName:String,
-   bvn:{
-    type:String,
-    required:true
-   },
-   productCategory:String
-}) ;
+        type:String,
+        required:true,
+        unique:true
+    },
+    
+    address:{
+        type:String,
+        lowercase:true,
+        required:true
 
-//creacting a model and exporting it
-const existingProposalModel = mongoose.model('existingProposalModel',existingProposal);
-module.exports = existingProposalModel;
+    },
+    dateOfBirth:{
+        type:Date,
+        required:true,
+        unique:true
+    },
+    stateOfOrigin:{
+        type:String,
+        required:true
+
+    },
+    BVN:{
+        type:Number,
+        required:true,
+        unique:true
+    },
+    phoneNumber:
+    {
+        type:Number,
+        required:true,
+        unique:true
+    },
+    NIN:{
+        type:Number,
+        required:true,
+        unique:true
+    },
+
+    ProductName:
+    {
+        type:String,
+    required:true},
+
+    productCategory:
+    {
+        type:String,
+        required:true
+    }
+  
+
+
+})
+
+const existingProposalModel=  mongoose.model('proposal not listed in the platform',structure2)
+
+module.exports=existingProposalModel
